requirements.txt
Flask
tensorflow
Keras
google-cloud-storage